Separated all existing admin module endpoints into their respective modules
