+++ src/modules/admin/admin.controller.js
@@ -94,250 +94,6 @@ const deleteAdmin = async (req, res, next) => {
   }
 };
 
-// --------------------- PATIENT CRUD ---------------------
-const getAllPatients = async (req, res, next) => {
-  try {
-    const patients = await adminService.getAllPatients();
-    res.status(200).json({
-      statusCode: 200,
-      message: "Patients retrieved successfully",
-      data: patients,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const getPatientById = async (req, res, next) => {
-  try {
-    const patient = await adminService.getPatientById(req.params.id);
-    if (!patient) {
-      const err = new Error("Patient not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Patient retrieved successfully",
-      data: patient,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const createPatient = async (req, res, next) => {
-  try {
-    const patient = await adminService.createPatient(req.body);
-    res.status(201).json({
-      statusCode: 201,
-      message: "Patient created successfully",
-      data: patient,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updatePatient = async (req, res, next) => {
-  try {
-    const patient = await adminService.updatePatient(req.params.id, req.body);
-    if (!patient) {
-      const err = new Error("Patient not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Patient updated successfully",
-      data: patient,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const deletePatient = async (req, res, next) => {
-  try {
-    const patient = await adminService.deletePatient(req.params.id);
-    if (!patient) {
-      const err = new Error("Patient not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Patient deleted successfully",
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-// --------------------- DOCTOR CRUD ---------------------
-const getAllDoctors = async (req, res, next) => {
-  try {
-    const doctors = await adminService.getAllDoctors();
-    res.status(200).json({
-      statusCode: 200,
-      message: "Doctors retrieved successfully",
-      data: doctors,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const getDoctorById = async (req, res, next) => {
-  try {
-    const doctor = await adminService.getDoctorById(req.params.id);
-    if (!doctor) {
-      const err = new Error("Doctor not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Doctor retrieved successfully",
-      data: doctor,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const createDoctor = async (req, res, next) => {
-  try {
-    const doctor = await adminService.createDoctor(req.body);
-    res.status(201).json({
-      statusCode: 201,
-      message: "Doctor created successfully",
-      data: doctor,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updateDoctor = async (req, res, next) => {
-  try {
-    const doctor = await adminService.updateDoctor(req.params.id, req.body);
-    if (!doctor) {
-      const err = new Error("Doctor not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Doctor updated successfully",
-      data: doctor,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const deleteDoctor = async (req, res, next) => {
-  try {
-    const doctor = await adminService.deleteDoctor(req.params.id);
-    if (!doctor) {
-      const err = new Error("Doctor not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Doctor deleted successfully",
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-// --------------------- MEDICINE CRUD ---------------------
-const getAllMedicines = async (req, res, next) => {
-  try {
-    const medicine = await adminService.getAllMedicines();
-    res.status(200).json({
-      statusCode: 200,
-      message: "Medicine retrieved successfully",
-      data: medicine,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const getMedicineByName = async (req, res, next) => {
-  try {
-    const medicine = await adminService.getMedicineByName(req.params.name);
-    if (!medicine) {
-      const err = new Error("Medicine not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Medicine retrieved succsessfully",
-      data: medicine,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const addMedicine = async (req, res, next) => {
-  try {
-    const medicine = await adminService.addMedicine(req.body);
-    console.log("Medicine added:", medicine);
-    res.status(201).json({
-      statusCode: 201,
-      message: "Medicine added successfully",
-      data: medicine,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updateMedicine = async (req, res, next) => {
-  try {
-    const medicine = await addMedicine.updateMedicine(
-      req.params.name,
-      req.body
-    );
-    if (!medicine) {
-      const err = new Error("Medicine not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Medicine updated succsessfully",
-      data: medicine,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const deleteMedicine = async (req, res, next) => {
-  try {
-    const medicine = await adminService.deleteMedicine(req.params.name);
-    if (!medicine) {
-      const err = new Error("Medicine not found");
-      err.status = 404;
-      return next(err);
-    }
-    res.status(200).json({
-      statusCode: 200,
-      message: "Medicine deleted successfully",
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
 module.exports = {
   getAllUsers,
   getAllAdmins,
@@ -345,19 +101,4 @@ module.exports = {
   createAdmin,
   updateAdmin,
   deleteAdmin,
-  getAllPatients,
-  getPatientById,
-  createPatient,
-  updatePatient,
-  deletePatient,
-  getAllDoctors,
-  getDoctorById,
-  createDoctor,
-  updateDoctor,
-  deleteDoctor,
-  getAllMedicines,
-  getMedicineByName,
-  addMedicine,
-  updateMedicine,
-  deleteMedicine,
 };
