+++ src/modules/admin/admin.service.js
@@ -101,272 +101,6 @@ const deleteAdmin = async (id) => {
   }
 };
 
-// --------------------- PATIENT ---------------------
-const getAllPatients = async () => {
-  return prisma.user.findMany({ where: { role: "PATIENT" } });
-};
-
-const getPatientById = async (id) => {
-  const patientId = id;
-
-  if (!patientId) {
-    const error = new Error("Invalid patient ID format");
-    error.status = 400;
-    throw error;
-  }
-
-  return prisma.user.findUnique({ where: { id: patientId } });
-};
-
-const createPatient = async (patientData) => {
-  const id = generateIdWithPrefix("P");
-
-  const { email, password, name, age, gender } = patientData;
-
-  if (!email || !password || !name || !age || !gender) {
-    const error = new Error(
-      "Missing required fields: email, password, name, age, gender"
-    );
-    error.status = 400;
-    throw error;
-  }
-
-  const hashedPassword = await bcrypt.hash(password, 10);
-
-  return prisma.user.create({
-    data: {
-      id,
-      role: "PATIENT",
-      email,
-      password: hashedPassword,
-      name,
-      age: parseInt(age, 10),
-      gender,
-    },
-  });
-};
-
-const updatePatient = async (id, patientData) => {
-  const patientId = id;
-
-  if (!patientId) {
-    const error = new Error("Invalid admin ID format");
-    error.status = 400;
-    throw error;
-  }
-
-  try {
-    return prisma.user.update({
-      where: { id: patientId },
-      data: {
-        ...patientData,
-        ...(patientData.age && { age: parseInt(patientData.age, 10) }),
-      },
-    });
-  } catch (error) {
-    if (error.code === "P2025") {
-      return null;
-    }
-    throw error;
-  }
-};
-
-const deletePatient = async (id) => {
-  const patientId = id;
-
-  if (!patientId) {
-    const error = new Error("Invalid patient ID format");
-    error.status = 400;
-    throw error;
-  }
-
-  try {
-    return prisma.user.delete({
-      where: { id: patientId },
-    });
-  } catch (error) {
-    if (error.code === "P2025") {
-      return null;
-    }
-    throw error;
-  }
-};
-
-// --------------------- DOCTOR ---------------------
-const getAllDoctors = async () => {
-  return prisma.user.findMany({ where: { role: "DOCTOR" } });
-};
-
-const getDoctorById = async (id) => {
-  const doctorId = id;
-
-  if (!doctorId) {
-    const error = new Error("Invalid doctor ID format");
-    error.status = 400;
-    throw error;
-  }
-
-  return prisma.user.findUnique({ where: { id: doctorId } });
-};
-
-const createDoctor = async (doctorData) => {
-  const id = generateIdWithPrefix("D");
-
-  const { email, password, name, age, gender } = doctorData;
-
-  if (!email || !password || !name || !age || !gender) {
-    const error = new Error(
-      "Missing required fields: email, password, name, age, gender"
-    );
-    error.status = 400;
-    throw error;
-  }
-
-  const hashedPassword = await bcrypt.hash(password, 10);
-
-  return prisma.user.create({
-    data: {
-      id,
-      role: "DOCTOR",
-      email,
-      password: hashedPassword,
-      name,
-      age: parseInt(age, 10),
-      gender,
-    },
-  });
-};
-
-const updateDoctor = async (id, doctorData) => {
-  const doctorId = id;
-
-  if (!doctorId) {
-    const error = new Error("Invalid admin ID format");
-    error.status = 400;
-    throw error;
-  }
-
-  try {
-    return prisma.user.update({
-      where: { id: doctorId },
-      data: {
-        ...doctorData,
-        ...(doctorData.age && { age: parseInt(doctorData.age, 10) }),
-      },
-    });
-  } catch (error) {
-    if (error.code === "P2025") {
-      return null;
-    }
-    throw error;
-  }
-};
-
-const deleteDoctor = async (id) => {
-  const doctorId = id;
-
-  if (!doctorId) {
-    const error = new Error("Invalid doctor ID format");
-    error.status = 400;
-    throw error;
-  }
-
-  try {
-    return prisma.user.delete({
-      where: { id: doctorId },
-    });
-  } catch (error) {
-    if (error.code === "P2025") {
-      return null;
-    }
-    throw error;
-  }
-};
-
-// --------------------- MEDICINE ---------------------
-const getAllMedicines = async () => {
-  return prisma.pharmacy.findMany();
-};
-
-const getMedicineByName = async (name) => {
-  const medicineName = name;
-
-  if (!medicineName) {
-    const error = new Error("Invalid Medicine Name");
-    error.status = 400;
-    throw error;
-  }
-
-  return prisma.pharmacy.findUnique({ where: { medicine_name: medicineName } });
-};
-
-const addMedicine = async (medicineData) => {
-  const { medicine_name, stock, price } = medicineData;
-
-  if (!medicine_name || !stock || !price) {
-    const error = new Error(
-      "Missing required fields: medicine_name, stock, price"
-    );
-    error.status = 400;
-    throw error;
-  }
-
-  return prisma.pharmacy.create({
-    data: {
-      medicine_name,
-      stock: parseInt(stock, 10),
-      price: parseInt(price, 10),
-    },
-  });
-};
-
-const updateMedicine = async (name, medicineData) => {
-  const medicineName = name;
-
-  if (!medicineName) {
-    const error = new Error("Invalid Medicine Name");
-    error.status = 400;
-    throw error;
-  }
-
-  try {
-    return prisma.pharmacy.update({
-      where: { medicine_name: medicineName },
-      data: {
-        ...medicineData,
-        ...(medicineData.stock && { stock: parseInt(medicineData.stock, 10) }),
-        ...(medicineData.price && { price: parseInt(medicineData.price, 10) }),
-      },
-    });
-  } catch (error) {
-    if (error.code === "P2025") {
-      return null;
-    }
-    throw error;
-  }
-};
-
-const deleteMedicine = async (name) => {
-  const medicineName = name;
-
-  if (!medicineName) {
-    const error = new Error("Invalid medicine name");
-    error.status = 400;
-    throw error;
-  }
-
-  try {
-    return prisma.pharmacy.delete({
-      where: { medicine_name: medicineName },
-    });
-  } catch (error) {
-    if (error.code === "P2025") {
-      return null;
-    }
-    throw error;
-  }
-};
-
 module.exports = {
   getAllUsers,
   getAllAdmins,
@@ -374,19 +108,4 @@ module.exports = {
   createAdmin,
   updateAdmin,
   deleteAdmin,
-  getAllPatients,
-  getPatientById,
-  createPatient,
-  updatePatient,
-  deletePatient,
-  getAllDoctors,
-  getDoctorById,
-  createDoctor,
-  updateDoctor,
-  deleteDoctor,
-  getAllMedicines,
-  getMedicineByName,
-  addMedicine,
-  updateMedicine,
-  deleteMedicine,
 };
