+++ src/modules/scheduling/scheduling.route.js
@@ -1,14 +1,15 @@
 // src/modules/admin/admin.route.js
 const express = require("express");
 const schedulingController = require("./scheduling.controller");
+const { authenticateToken, authorizeRoles } = require("../../middlewares/auth.middleware");
 
 const router = express.Router();
 
 // CRUD SCHEDULE
-router.post("/createSchedule", schedulingController.createSchedule);
-router.get("/getAllSchedules", schedulingController.getAllSchedules);
-router.get("/getScheduleById/:id", schedulingController.getScheduleById);
-router.put("/updateSchedule/:id", schedulingController.updateSchedule);
-router.delete("/deleteSchedule/:id", schedulingController.deleteSchedule);
+router.post("/createSchedule", authenticateToken, authorizeRoles("ADMIN", "DOCTOR"), schedulingController.createSchedule);
+router.get("/getAllSchedules", authenticateToken, authorizeRoles("ADMIN", "DOCTOR"), schedulingController.getAllSchedules);
+router.get("/getScheduleById/:id", authenticateToken, authorizeRoles("ADMIN", "DOCTOR", "PATIENT"), schedulingController.getScheduleById);
+router.put("/updateSchedule/:id", authenticateToken, authorizeRoles("ADMIN", "DOCTOR"), schedulingController.updateSchedule);
+router.delete("/deleteSchedule/:id", authenticateToken, authorizeRoles("ADMIN", "DOCTOR"), schedulingController.deleteSchedule);
 
 module.exports = router;
